Added some experimental functionality.
Added function to downsample objects from a large data source, making
reduced CSV of data points redundant (also deleted this file from
repository). Added a new gradient method to calculate the impact of
repositioning an object on the global reconstruction error (not just
the error for the object in question).